requirements.txt
fastapi
uvicorn
pydantic-ai
pydantic-ai-slim[openai]
devtools